requirements.txt
streamlit>=1.30.0
openai>=1.10.0
chromadb>=0.4.22
PyPDF2>=3.0.0
numpy>=1.24.0
python-dotenv>=1.0.0